Worker status to attachment status
