+++ libs/pandautils/src/main/java/com/instructure/pandautils/features/inbox/utils/InboxMessageView.kt
@@ -14,7 +14,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package com.instructure.pandautils.features.inbox.util
+package com.instructure.pandautils.features.inbox.utils
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
